+++ src/main/java/org/getspout/vanilla/protocol/codec/SetWindowSlotsCodec.java
@@ -50,7 +50,7 @@ public final class SetWindowSlotsCodec extends MessageCodec<SetWindowSlotsMessag
 				if (ChannelBufferUtils.hasNbtData(item)) {
 					nbtData = ChannelBufferUtils.readCompound(buffer);
 				}
-				items[slot] = new ItemStack(MaterialData.getMaterial(id, (short) damage), itemCount, (short) damage).setAuxData(nbtData);
+				items[slot] = new ItemStack(MaterialData.getMaterial((short)id, (byte) damage), itemCount, (short) damage).setAuxData(nbtData);
 			}
 		}
 		return new SetWindowSlotsMessage(id, items);
